Infer bonus from bonus estimated time.
This commit adds `texpy.commands.get_bonus` as an analog to
`get_rewards`. It will try to infer a Bonus given an estimated bonus
time and the current reward rate.

This commit also fixes a minor bug that used to throw a `KeyError` for
tasks that did not provide a `Bonus` field.