Added task-wide mean_aggrement as a metric
This commit adds a new metric to compute aggrement with the majority
label for a task. The existing `texpy.metrics.mean_agreement` actually
computed mean-agreement for each worker: this commit renames it to
`texpy.metrics.mean_agreement_per_worker` to make that clearer.

This commit also renames `simple_agreement` to `pairwise_agreement` to
be more informative for users.